requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
reportlab==4.0.4
python-dotenv==1.0.0
gunicorn==20.1.0